+++ com.unity.ml-agents/Runtime/Policy/HeuristicPolicy.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System;
+using MLAgents.Sensors;
 
 namespace MLAgents
 {